Altered the console to set the canonical hostname rather than an alias
The /etc/hosts man page describes a difference between the "canonical_hostname"
(fqdn) and the aliases for that hostname.

This difference is reflected in the behaviour of some external tools such
as cloud-init and freeipa.
These tools will retrieve the hostname (typically via the hostname
command or from /etc/hostname) then consult /etc/hosts to find the
fqdn by taking the name in the "canonical_hostname" place in the line
where the found hostname is an alias.

This can cause problems when the application acts differently based on
whether the fqdn is "localhost" or not.

https://bugzilla.redhat.com/show_bug.cgi?id=1286830


(transferred from ManageIQ/manageiq-gems-pending@e28b319c1a883d98ded3d551502cc3fc7b66a52f)