Prevent Trollop from exiting the rspec process
Trollop.die and Trollop.educate exit the calling process with a 0
exit status.

This is undesirable as it will cause the spec suite to go green without
having run all the specs. In addition, the writer of a spec that would
cause the process to exit likely did not want that behavior and would
expect the spec to fail.

This change achieves that behavior by stubbing the calls to Trollop
.die and Trollop.educate to raise errors.

Now we can also test that these methods are called by expecting
that these particular error classes are raised.


(transferred from ManageIQ/manageiq-gems-pending@4017e44ebc56a65630bec4de83354dc10ddcd700)