Look for translation files under Rails root directory in addition to to inside the gem
Now that gems/pending is a real gem, the translation file for productization is no longer in a path relative
to the ruby script that loads it. That file is in the original location under the Rail root directory. We need to load files
from that location as well.

https://bugzilla.redhat.com/show_bug.cgi?id=1422433


(transferred from ManageIQ/manageiq-gems-pending@7d12a12f3f0e482629ccab1a5128743310425b42)