Update SCAP rules yaml file path to new location under productization directory.
Now that gems/pending is a real gem, productized file is no longer in a path relative to the ruby script that loads it.
The file will now created under the productization directory during the build. We need to point to that location to load the file.

https://bugzilla.redhat.com/show_bug.cgi?id=1422433


(transferred from ManageIQ/manageiq-gems-pending@b8fe0cee968de9d71b385c101aabc0372a395dcd)