Extract Networking Settings to a submenu
There are already 4 options for network configuration. With IPv6 this
will be extended.

Logically these 4 options are tight together and they are unrelated
from other options. Once the user is done with networking, he or she may
not want to return back.

I have been thinking about this and I wanted to avoid this move for a
while. Then, I have come to a conclusion that this should pay off in
the long run. There is growing ammount of options in the
appliance_console and separation will make it easier for mammals to
comprehend.


(transferred from ManageIQ/manageiq-gems-pending@62a98c34351e11c5864dd47c823d721022b6c7b7)