Check for existing nodes before attempting registration
This implements a check during the "ask questions" phase of setting
up a standby server which will inform the user if they are trying to
use a node number which is already assigned to a node record.

Before this change the initial clone and replication setup would
succeed, but the registration would fail if a user tried to re-use
a node number (from a previously failed node, for example). This
would leave you in a state where the new standby server is started
and is replicating from a new replication slot on the primary, but
cannot join the repmgr cluster.

This can be avoided by registering with the --force flag, but we
don't really want to do that unconditionally so we have to connect
to the primary to determine if the requested node number is in
use *before* doing the clone then prompt the user for confirmation.


(transferred from ManageIQ/manageiq-gems-pending@0e16d08a7fba4f44aa8af59693ef7b043d7867ba)