Stop the postgres and repmgr services before configuring a standby
This ensures that we will not error out in the "reconfigure" case.

In this case the standby server is configured successfully and we
go through the process of configuring a standby server again. Before
this change we would fail because the running PG server would be
using the replication slot that we wanted to replace.

This error would leave the primary server with a replication slot
and would remove the data from the standby server.

https://bugzilla.redhat.com/show_bug.cgi?id=1445841


(transferred from ManageIQ/manageiq-gems-pending@be284fee6c0e861889556378f91855f954c44c4c)