Choose restore type based on the target file type
This commit allows us to support both the new, pg_basebackup
method of backup as well as the old pg_dump method.

The pg_basebackup output is just a tgz file of the contents of
the data directory, so restore is just replacing the current data
directory with the data in the archive.

https://bugzilla.redhat.com/show_bug.cgi?id=1495192


(transferred from ManageIQ/manageiq-gems-pending@dbb78da63c34a01f4a5be0a4c53fe6f6830a7573)