Handle multi-value args in PostgresAdmin.backup_pg_dump
There are a few args passable to `pg_dump` that can take multiple
values.  Since it is safer (in my eyes) to do `-t table1 -t table2`
instead of trying our luck with `-t table1 table2`, this change makes it
so the options passed to AwesomeSpawn do the former and not the latter.

To accomplish this, this means we have to move the arguments outside of
the `opts` hash, and into the `args` array.

Also added a pretty decent amount of tests around this to make sure it
works as expected.


(transferred from ManageIQ/manageiq-gems-pending@156e1ec330df8b940872be93c727c61cf9ea9cc4)