Allow #ask_yn? to be modified via block
The `#ask_yn?` method is one of the few methods in `Prompts` that does
not make use of the `just_ask` question.  `#just_ask`, by default, has
everything flow through `Readline`, which has a different output than
the standard commands.

Because of this, if you try mixing the two in the specs, you can get
some strange behavior, such

    EOFError: The input stream is exhausted.

Which comes from the Highline lib.  This happens because the input is
trying to be read from two different sources at once, and doesn't
function as expected doing that (one ends up being blank, even though
some input is still available).

* * *

Instead of making the risky change of forcing `q.readline` for every
usage of `ask_yn?`, allow for a block, so this can be done a-la-carte as
needed.