Adds DatabaseAdmin#ask_for_tables_to_exclude_in_dump
Adds a set of questions to be queried for `:dump` based actions in
DatabaseAdmin.  It asks:

  - If any tables wish to be excluded
  - If yes, what tables should be excluded

The defaults for this are set as some sensible and standard defaults to
be used, and allow for quickly moving on, but no tables will be excluded
if the user answers no to the first question.