Don't add the schema to the search path
The new schema is no-longer dependent on the cluster name (as it was
removed), it is now always "repmgr". Additionally the tables names
are much more generic and more likely to conflict with other tables
(e.g. nodes, events).

This commit also adds the new schema name to all places which
previously referenced the repmgr tables.