Write the pgpass file earlier
This will allow the password to be used when repmgr is checking if
the new record is sync-ed to the standby. Previously it was failing
to connect because it was not using the PGPASSWORD env var provided
with the `standby register` command.

Either way writing this file earlier can't hurt.