Fix bad validator for Backup/Dump commands
DatabaseAdmin::LOCAL_FILE_VALIDATOR was originally written to support
the `:restore` action, which did indeed need a file to exist for it to
work properly.

The `:backup` and `:dump` actions, however, were much more likely that
they wouldn't have an existing file, so this was not a valid validator
for those two actions, and causes the user to end up in an infinite loop
in the console.

(while this could still work with this validator in place as is, it
would require having to create a dummy dump/backup file prior to running
the command, which is clearly not obvious and a poor user experience)