Confirm `DatabaseAdmin#ask_*` methods return true
Updates all of the specs, and makes a slight code adjustment to
DatabaseAdmin to confirm that all `ask_*` methods return true.

Reason being is that in the `bin/appliance_console`, their is the
following common practice:

    db_admin.ask_questions && db_admin.activate

This means that if a method returns false, it won't move on to execture
the `db_admin` command.

Since all of these questions should be returning truthy in each of these
cases, the specs are there so that they confirm this behavior incase the
order of the questions is ever changed in `DatabaseAdmin.ask_questions`