Fixes rake arguments for ask_nfs_file_options
Currently users were limited to only selecting the NFS share URL for
dumps and backups, and only the `restore` would work as expected since
it was the basis for the DatabaseAdmin class when it was created.

This change will maintain the original changes for the restore actions,
but then first ask for the name for the filename when doing a dump or
backup, and then the URI endpoint for the mount for the NFS share.