Fixes signal trapping
Namespacing the error class allows the trap to work as expected again.

* * *

Background:

When this was initially extracted from the `manageiq-gems-pending` repo,
the lib/manageiq/appliance_console/errors.rb looked like this:

    class MiqSignalError < RuntimeError; end

Link to the PR and removed file:

- https://github.com/ManageIQ/manageiq-gems-pending/pull/292
- https://github.com/ManageIQ/manageiq-gems-pending/pull/292/files#diff-4265a66ff1f90898eb4add59adb4c5ca

During the initial commits of the newly created repository, the
errors.rb was namespaced:

https://github.com/ManageIQ/manageiq-appliance_console/commit/ff0dea9e#diff-30f4d0543374373ec9d55b71f7ce6301

But the `bin/appliance_console` in the that commit was not updated to
reflect that.  Since the specs don't test against the
`bin/appliance_console`, this was never noticed since `v1.0.0` of the
gem.