Use io/console instead of `stty`
Since ruby-1.9.3, 'io/console' has been a part of ruby's stdlib:

    https://svn.ruby-lang.org/repos/ruby/tags/v1_9_3_0/NEWS

This adds the `STDIN.noecho` when `io/console` is loaded, which
effectively does the same as what we were doing with system calls, with
the addition of also handling `signals` properly (which we were not
doing, and would have needed the `isig` flag set to do so).

Combined with the block form of `STDIN.getc`, we can replicate what was
being done previously, and avoid multiple system calls every time
`press_any_key` was called.