Adds specs for #ask_s3_file_options
Had to move the `region` question up since there is weirdness with the
question input when something that doesn't use readline is used before
readline questions.

Pretty sure it is `readline` slurping up all of the options, which works
for it, but then doesn't leave anything behind afterwards for others...
regardless, something I will fix another time, and this makes the tests
work as they did with the others.

Pending specs will be addressed in the next commit, when I add support
for the `:remote_file_name` argument for s3.