Make s3 uri paths optional
Based on the format provided in the sample URL, and how the
implementation is for the MiqS3Session, the bucket name is plucked from
the host.  In the case of s3, if no path is provided to the URL, it
means that the resulting directory is top level, which is should be
valid.

This fix simply allows that functionality by updating the prompt
validator to accept no paths in the uri conditionally (based on an
argument option) and hooks up the `#ask_s3_file_options` method to
use this new flag.  All other uses of `#ask_for_uri` should still
function as they did.