Revert to using --xlog-method fetch for postgres backup
This is because we cannot use the tar option with stream.
We need the backup to be written in a single stream for the work
regarding file splitting and backing up to cloud-based targets to
move forward.

Fixes #366


(transferred from ManageIQ/manageiq-gems-pending@640f464d1d22626e5522b48397f8623c0ae07904)