Handle pipes for pg_restore
For whatever the reason, pg_dump doesn't like reading from a fifo when
it isn't passed in as STDIN.

This assigns the same treatment to it as was done to pg_dump/pg_restore,
which is passes it through STDIN.  Since we don't really have good
access to STDIN via `AwesomeSpawn`, we use vanilla `Kernel.spawn` to
handle the actually command execution.  We call out to `AwesomeSpawn`
for the command builder, as well as emulate the errors that it would
throw so there is limited change in how this library behaves to the
user.


(transferred from ManageIQ/manageiq-gems-pending@912ccfbd1f4c8cb5c25e5e216a1ee7e4346855ec)