Add postgresql-9.5 to travis
Allows us to actually test things with a running postgres installation.

Also updates the rake tasks to differentiate things further:

- `rake spec`:     Runs all specs minus ones that require pg running
- `rake spec:ci`:  Specs run in a travis environment
- `rake spec:dev`: Specs run in a development environment

The difference between `spec:ci` and `spec:dev` is the `spec:dev` will
load in the `PostgresRunner` lib, and CI is configured to work with the
postgresql addon from travis and its specific setup/config.

Since CI is ephemeral, though, we don't have to worry as much above
trashing a dev machine's current database when doing a restore, so this
is mostly just making sure we are restoring our configs and data in the
right spot, and triggering start/stop commands in the way the service on
CI expects.

The other thing we do have to worry about, however, is that
PostgresAdmin assumes being root when doing it's restore.  That isn't
really easy to do in the same process on CI, so instead a subprocess is
called that loads in the minimum libs necessary for PostgresAdmin, and
calls `PostgresAdmin.restore` with elevated permissions (aka:  `sudo`)


(transferred from ManageIQ/manageiq-gems-pending@11bd1743a5adbb66d7a9b4d022e37081d56e21d6)