Add PostgresRunner as a helper
This class is meant to be in charge of spinning up postgres on
environments where an existing postgres instance already exists, and
blowing away the data directory to run `.restore_pg_basebackup` is
unacceptable.

It leverages the `pg_ctl` command to handle `initdb`, and starting and
stopping the instance.

It also tries to maintain the same interface as a `LinuxAdmin::Service`
instance so it can be used to stub that out in the specs as well.

Also, it will currently fixup and manageiq config includes in the
`postgresql.conf` file that are brought in by the backup.


(transferred from ManageIQ/manageiq-gems-pending@0b320df4733dd4a91efd05806c196cffeefb4c2e)