Simplify Filename Prompting
Simplify processing for the filename prompts.
The previous incarnation was called out as overly complex by CodeClimate.

In addition, due to a previous PR some of the backup_type constants had been
removed in favor of simple methods so use the #local_backup? when appropriate,
and add #object_store_backup? instead of constants for Swift and S3.