[PostgresAdmin] Fix backup_type being ignored
With the previous implementation, the code would always attempt to find
the magic number of the file for a `pg_dump` before checking if the
`backup_type` value was instead a `:basebackup`.  When streaming a file,
this would cause it to fail farther down the line since it would have
strip some of the content off from the beginning of the file,
invalidating it as a proper backup.

This change fixes by forcing the check for the backup type first to
avoid reading the file data if possible, and then by inspecting the file
contents when as a backup option.


(transferred from ManageIQ/manageiq-gems-pending@5b9e01790cf8cb43123cc59bd8c5739ae6e7c903)