Fix permissions on PG user home directory after disk mount
When we moved from SCL to normal PG the data directory moved
into the postgres user's home directory (/var/lib/pgsql). We use
the directory over the data directory as a mount point for the
database filesystem.

Mounting the filesystem on the home directory causes permissions
issues during initdb as the postgres user expects to be able to
write into it's home directory. This PR fixes the permissions
when we mount on /var/lib/pgsql