Handle the Optimist::Parser instance methods that exit
This is needed to detect when Optimist may exit the spec process
from inside the parser instance. This would cause a "green" status
when in reality one of the specs caused the entire process to exit.

This means that only a subset of the specs would run, but no failures
would be reported.

In particular this can happen if you provide an option to the Cli
parse command that is not in the options list.