Point this version to manageiq-gems-pending@hammer
This still has the util/miq-password file that we are requiring.
Because this dependency is declared in the Gemfile rather than
the gemspec we will end up using whatever version is bundled with
the rest of the application.

On versions of the appliance that use this version of the console,
that's hammer.