[Utilities] Add #disk_usage
As one of two things that is still making use of `manageiq-gems-pending`
(the other being `PostgresAdmin`), this only a single use method that
doesn't need to bring then entirety of `manageiq-gems-pending` along
with it in the build.

For `#disk_usage`, porting over the singular method itself, while
removing the `linux` vs `mac` differences (only keeping the linux),
should be good enough for the use case required by the appliance
console.