[spec/] Avoid ENV overwrite in specs
The APPLIANCE_PG_DATA and APPLIANCE_PG_SERVICE environment variables are
used by both the database_replication_spec.rb and postgres_admin_spec.rb
specs.

The previous implementation, specs in both database_replication_spec.rb
and postgres_admin_spec.rb would delete/overwrite the ENV variables that
were assumed to be set for the other to run it's processes.  In this
case, it is simpler to store the variable in a tmp variable before
overwriting it for the purposes of the spec in question, and then reset
the var to it's original value when the spec is finished.  That way
other specs will be able to use it with out issue, regardless of what
order the specs are triggered from.