[ci_helper.rb][CI] Set ENV vars earlier
Many specs will make use of `LinuxAdmin::Service`, which is overwritten
in `spec/ci_helper.rb` to make use of `CiPostgresRunner` if the service
in question is `postgres`.

However, since calling the overwritten method requires the
`.service_name` method:

    module LinuxAdmin
      def Service.new(service_name)
        if ManageIQ::ApplianceConsole::PostgresAdmin.service_name == service_name
          CiPostgresRunner
        else
          super
        end
      end
    end

That in turn requires `ENV["APPLIANCE_PG_SERVICE"]` to be set, which it
is not.

The simplest fix for this case is to just make sure the environment
variables are loaded in CI where we need them.  Previously, they were
only needed when launching the sub process, but now with other specs
making use of `LinuxAdmin`, it is also required to be set all the time.

There was also a slight readability fix with the `system` call to remove
the temporary variable, to make it more clear what were helper vars, and
what was the command being run.