[DatabaseAdmin] Use PostgresAdmin directly
Instead of calling a rake task in core, call `PostgresAdmin` which is
now part of this repository.  This takes many of the checks that exist
in `lib/evm_database_ops.rb` (core) and puts them in this class,
tweaking them slightly so they don't require things like `ActiveRecord`
to function, or classes that only exist in core
`(VmdbDatabaseConnection` for example)

In addition, this also doesn't include any of the `.with_file_storage`
calls as that will be phased out.  `PostgresAdmin` just needs a local
file to write to, and since `MiqFileStorage` used a FIFO to write data
to external sources (s3, swift, FTP), there is no change required there,
and this can just accept a local file and pass that to `PostgresAdmin`.