[PostgresAdmin] Use pg-logical_replication
Use PG::LogicalReplication::Client to remove subscriptions or
publications on a db prior to doing a database reload.

Doesn't drop the `MiqRegion`, as was previously done as part of
PglogicalSubscription in core:

  https://github.com/ManageIQ/manageiq/blob/8ca93da126aada556685db5d05aaf5e9a0666900/app/models/pglogical_subscription.rb#L70

But since this is a database reload, and the Region will be removed from
the DB anyway, this should not matter, and the necessary code required
to perform that delete will be pointless to include in this gem.