[PostgresRunner] Add .hard_reset (local only)
Adds a helper method to PostgresRunner (.hard_reset), which is used to
reset the database to a fresh state.

Required for local tests only since the DB backup was taken off of an
appliance, so when that is restored first, it changes the default DB
roles to that `.initial_setup` create and is assumed by the class.