[postgres_admin_spec.rb] Ensure env is setup
For the `.restore` specs, ensure the env is setup properly so the specs
can run properly.

With the new methods that check the existing database connections in
`PostgresAdmin` directly, this allows said connection to work properly.

For CI, we don't want to do this, since it doesn't actually setup the
ENV (bad method name, my bad...), but runs the restore in a task all by
itself.