[PostgresRunner] Simplify .restore and specs
Previously, the logic to pre-determine the backup type required
knowledge of the file storage type that PostgresAdmin did not have.  In
the case where the file storage data was being streamed (S3, FTP, etc.)
and was only read from the source once, you had to call out to the
`PostgresAdmin.pg_dump_file?` and `PostgresAdmin.base_backup_file?`
calls manually to determine what type file was going to be restored so
the proper code could be executed, and `PostgresAdmin` didn't have to
worry if the file data was being streamed or was just read off of a
disk/mount.

Since .validate_backup_file_type (not show in diff) is now part of
.restore, this extra case logic is no longer necessary, as `backup_type`
will always be set if not passed in.

As such, the code and spec for it has been updated to reflect that.