Only start evmserverd after db+messaging configured
Currently when configuring the application evmserverd is started after
setting up the database, it is then restarted twice after configuring
messaging (once for unconfiguring then once after configuring).

This is very inefficient and when we move to require messaging to be
configured in order to start evmserverd then the first start won't ever
complete anyway.