Handle default value output in highline
If we enter a value that is not a default value, but one is derived, then the actual value entered is output.

so if we type in text that infers option 1, then highline now outputs (1)

Not exactly sure why there is a change for a default value being presented,
I'm guessing the value was not checked by the previous test. So the change in expect_heard
shows this at this time