Both pg_dump and pg_basebackup are supported ways to backup the database
There are pros and cons to both logical and physical backups and in fact
people should probably plan on doing both types of backups.

Physical/pg_dump for table or database exports for testing, upgrades, and for
disaster recovery. It's slower but generally smaller.

Logical/pg_basebackup for cluster backups and versatility to restore from point
in time or from standbys. It's also much faster.

See also:
https://www.enterprisedb.com/postgresql-database-backup-recovery-what-works-wal-pitr