Remove ping of IPA Server
An IPA server doesn't have to be pingable in order to be valid.
Specifically, the FreeIPA demo server doesn't repond to ping requests,
and so can't be used with this code in place.

We are going to try to connect to the server during setup almost
immediately after, and it will fail with a network error anyway, which
is typically more descriptive.